services/lora-training/app.py
"""
FastAPI service for LoRA training worker
Processes training jobs from the API Gateway
"""

from fastapi import FastAPI, HTTPException, BackgroundTasks
from pydantic import BaseModel, Field
from typing import Optional, List
import os
from datetime import datetime
from dotenv import load_dotenv
import asyncio

# Load environment from root
load_dotenv(dotenv_path='../../.env')

app = FastAPI(
    title="LoRA Training Service",
    description="Worker service for processing LoRA training jobs",
    version="1.0.0",
    docs_url="/docs",
    redoc_url="/redoc"
)

# Pydantic models for request/response
class TrainRequest(BaseModel):
    """Request model for training a LoRA"""
    job_id: str = Field(..., description="Job ID from MongoDB")
    user_id: str = Field(..., description="User ID")
    video_url: str = Field(..., description="Source video URL (HTTP/HTTPS or S3)")
    lora_name: str = Field(..., description="Name for the LoRA model")
    trigger: str = Field(default="person", description="Trigger phrase")
    steps: int = Field(default=2500, ge=1000, le=10000, description="Training steps")
    learning_rate: float = Field(default=0.00009, description="Learning rate")

class TrainResponse(BaseModel):
    """Response model for training request"""
    job_id: str
    status: str
    message: str

class HealthResponse(BaseModel):
    """Health check response"""
    status: str
    service: str
    version: str
    timestamp: str

@app.get("/health", response_model=HealthResponse, tags=["System"])
async def health_check():
    """Health check endpoint"""
    return {
        "status": "healthy",
        "service": "lora-training-worker",
        "version": "1.0.0",
        "timestamp": datetime.utcnow().isoformat()
    }

@app.post("/train", response_model=TrainResponse, tags=["Training"])
async def train_lora(request: TrainRequest, background_tasks: BackgroundTasks):
    """
    Process a LoRA training job

    This endpoint receives a training request and processes it in the background.
    Updates MongoDB job status as training progresses.
    Uploads final model to S3.
    """
    try:
        # Add training task to background
        background_tasks.add_task(process_training_job, request)

        return {
            "job_id": request.job_id,
            "status": "processing",
            "message": "Training started in background"
        }
    except Exception as e:
        raise HTTPException(status_code=500, detail=str(e))

async def process_training_job(request: TrainRequest):
    """
    Background task to process training job

    Steps:
    1. Update MongoDB status to 'processing'
    2. Download video from URL/S3
    3. Extract frames using existing pipeline
    4. Build dataset
    5. Upload dataset to S3
    6. Train via fal.ai
    7. Download trained model
    8. Upload to S3
    9. Update MongoDB with results
    """
    # TODO: Implement full training pipeline
    # For now, this is a placeholder

    print(f"Processing job {request.job_id}")
    print(f"Video: {request.video_url}")
    print(f"LoRA name: {request.lora_name}")

    # Simulate work
    await asyncio.sleep(2)

    print(f"Job {request.job_id} completed (placeholder)")

# Development server
if __name__ == "__main__":
    import uvicorn
    port = int(os.getenv("LORA_TRAINING_PORT", 5001))
    uvicorn.run(
        "app:app",
        host="0.0.0.0",
        port=port,
        reload=True
    )
